C is sooooo retarded.
I made a bunch of macros to do double arithmetic, and of course the code
that is generated is a joke. I'd ____REALLY____ prefer to stick with a
small handful of easy-to-write and easy-to-read, and efficient, asm
routines than a boatload of FUCKED UP #define macros that STILL don't say
what I want because the designers of C were TOO RETARDED to realize
programmers of SYSTEMS LEVEL SOFTWARE might ACTUALLY want access to the
CARRY bit, and that by making their LAME decisions they have sentenced
scads of C programmers to a semi-hell of tangled code.

We can, and will, break this cycle of pain: by writing tiny amounts of
assembler code that say, neatly, what we want to say.

Sorry about the SHOUTING, but I'm feeling rather strongly about this.


git-svn-id: file:///home/david/tashtego/svn/trunk/muforth@197 f1bc7e2d-46f1-0310-96be-f91141564092