A new compiler is checked in.  This version uses an abstract opcode set with a machine-independent interpreter.  In theory, this will make porting muForth to other processor architectures no work: muForth should just run on any UNIX like environment.
In the process of building this compiler, a few things were learned:

1) Because the target code is no longer the native processor's code,
calling code whose address has been placed in a vector must be through
a routine which will determine if the target code is C compiled code
or Forth compiled code.  If that latter, then the interpreter must be
invoked.

2) 'execute' really is processor specific, so that has been moved to
the compiler.  (As will be 'jump', but right now that's just removed.)

3) :entry isn't needed any more as the compiler can count the number
of paths of execution it is currently generating code for.  When that
count is 0 and new code is generated, then the count gets incremented.
At that point if the processor needs to do anything at the beginning
of a function, it may.

4) There is a library of math routines that should work on any processor.


git-svn-id: file:///home/david/tashtego/svn/trunk/muforth@218 f1bc7e2d-46f1-0310-96be-f91141564092