I added a games and lib directory.  I didn't add everything.  And, I may have split it improperly.  Also, I did -not- delete the original files.  We should do that as soon as we agree on this idea.
Further, I think a dev folder would be useful for device drivers.  A
series of pseudo devices is next on my list of things to write.  For
example, an X11 application that simulates an 2x16 char LCD or a PALM
bit-mapped LCD would make development of software on a host machine easier.

And, I can see pseudo devices for EEPROM's, keypads, keyboards, serial
output, etc.  But, now I'm just capturing notes in the wrong place.  :-)


git-svn-id: file:///home/david/tashtego/svn/trunk/muforth@221 f1bc7e2d-46f1-0310-96be-f91141564092