muforth banner no longer prints svn's revision number
Now that muforth is hosted in git rather than Subversion, I've started to
make some necessary changes. At some point I'll probably remove all the $Id$
keywords but for now they they are ugly but don't cause trouble.

However, the way that I parsed the revision of startup.mu4 and printed it at
startup breaks unless svn has checked out the code (and rewritten the
$Revision$ to "$Revision: <num> $". Of course, git isn't going to do this,
so if you do a clean checkout from git, build, and run, it barfs.

But not any more. Now it prints nothing! I tried having it print the first
16 chars of the commit sha-1 hash to identify itself, but, while nice, this
causes problems. When do we copy the commit sha-1 to version.h? Every time
we build? It gets annoying. So I left it out entirely. The build date is
still printed, and if it's truly necessary to know what commit a given build
is based on, it's easy to type "git show".

First real change via git! Yahoo!