New copyright, new license, new homepage link
Touched almost every file! Lotsa sed magic. Now the file headers are much
less cluttered. All is much better, and I'm checking in scripts (in
scripts/) to make editing the project homepage or copyright easier in
the future.

I switched from Apache license to MIT license, and changed the homepage
to point to an as-yet-non-existent page. That comes next. ;-)