target/ARM/ is working again, for AEB-1 target board
This isn't *that* exciting, since I first got it all working two years
ago, but it's working *again*, which is nice.

I cheated wherever possible, but the board really is running a headless
Forth. There is no way to do incremental compilation, and I leave the
board init (mostly) to the onboard boot-loader. But still. It works.