An almost complete checkin of AVR assembler
Condition code support is in, but no conditional branching yet.

I wanted to check this in because I tried a kind of postfix indexing
syntax. I liked it until I realized that it puts the offset (in the case
of ,y and ,z) very far from the index register (the ,y or ,z).

Syntax is (using "ld" as an example):

  <off> <reg> ld ,y

Without an offset it reads nicely:

  <reg> st z+

I'm going to try putting the index ops before the instruction and see what
that looks like. Something like this:

  <off> ,y 16 ld