Beginnings of loader & memory map for target/AVR
The loader is just a load file to bring in all the bits we need in
order to do target compilation. The "memory map" is words to create
an empty image (for the program memory), store into it, fetch from it,
etc. Pretty simple. This is where we capture the endianness of a target.