Fixed a subtle bug with "latest"
It used to be defined as "current @ @" - ie, get the value of current -
(which points to the dict chain that we're compiling onto) and dereference
it. This yields a pointer to the last word on the chain that current points
to, but it might not actually be the last word defined! Here's how:

: label  here target constant forth  asm ;
: code   label  does>  rcall ;

Translation of label: grab the current address in our target memory image
(here), switch to target chain, make a constant, switch back to forth, then
run the assembler.

Translation of code: make a label, then change its runtime to compile
itself (ie to compile a call to its label).

When the "does>" in "code" runs it calls "latest" to get the address of the
latest word define so it can rewrite its "high-level" code pointer.
This should be the "code" word we just defined, but will instead be
whatever was last defined on the forth chain. Bad. In fact, this code
segfaults. Eventually.

Maybe switching to forth aftwards is pointless; but this highlights the
fragility of this approach.

Now, when a new word is added to the dict, "latest" captures a pointer to
it. Switching current around has no effect on the value of "latest". With
this change, the above meta-compiling code example works fine!