Fixed copyright - again! - to more accurately reflect reality
Having done some source code archaeology today, I realized that the
existing copyright notice was rather optimistic about how long I've
been working on muFORTH. While a few bits - like the original ARM
assembler - there are two in the tree! - dates back to 1998 or so,
none of the core of muFORTH seems to date back before 2002.

There are in fact lots of files - eg in the target compiler code -
that say things like

   ( 2002-mar-29. Converting to my new Forth, muForth.)

A second corroboration is the date of the first commit (originally
into CVS) in Sept 2002.

Having forgotten the timeline, I'm guessing I wrote muFORTH in early
2002 and checked it into CVS that September. It makes me feel better
that I have not been working on it for 11 years.. it doesn't feel like
that long. I can believe six years. ;-)