Removed (literal) from C kernel and added literal; startup.mu4 adjusted
A small change, mostly because I wanted to experiment, but also because
I was annoyed with the first few lines of startup.mu4. Initially I
wanted to add ";" back into the kernel so startup would read better, but
this meant defining lots more things in C than I wanted. Keeping ^ and [
in the kernel is fine.

But having \[ (basically [ but defined in .forth. so it's easier to
compile, before we have \ or \c) was ugly. I removed it, and figured out
a way to compile [ from .compiler. - simply by calling the necessary
words:  token [ .compiler. find . We have a "live" interpreter for a
reason!

I started to look at taking : out of the C kernel. It would be easy
actually. I may still. Just to make a point.

Then I started thinking: how deep could you go? I realized that as long
as you had a way, by writing in memory, to affect your environment, that
the initial loader could simply be a text interpreter loop that looks up
tokens and executes them. The sequence of tokens in startup.mu4 would be
a "linearized" history of the call sequence of actually the Forth code.
This could be generated by a special-purpose compiler... Again, and
interesting rabbit hole, but not one I'm prepared (at the moment) to
descend into.

So take this checkin as my giving the nod to this idea.