As threatened, I've removed : from the C kernel.
This is really just for fun. It doesn't do anything differently. But now
the C kernel starts not knowing : or ; . The startup code defines first
: and then, after a bit of futzing around with dictionary chains, ; .
After that things are normal.

If there is a point to this change, it's to show that the C kernel
contains everything but the very last shell layer. Defining : and ; in
Forth is a neat trick, but not a deep one. What would be more impressive
would be to do this given only bare C primitives - ie, no words defined
in C that mimic Forth words by simply calling a sequence of primitives.
That would be more interesting.

And this idea gives rise to another bootstrap idea. Build a much simpler
and more monolithic C kernel, leaving only the lowest-level pieces
accessible to Forth. At startup, we would rebuild our own versions of
the compiler and interpreter, and throw away the original C ones (by not
using them - they would still be sitting around taking up space, but
only a few K).