Spring cleaning 2: literal and branch compilers are in Forth
I've given up on supporting both ITC and native. Or if I do still want
to do that, I'll have a flag that startup.mu4 can test, and build one or
the other. But the cruft I had to put into engine-itc.c to support the
"generic" control structures annoyed me. It's gone.

configure.sh and Makefile.in have no idea about anything other than
engine-itc.

(lit) is in C but literal is in Forth.

And I defined fig-FORTH's compile. ;-)