Added Vim settings for editing muFORTH files
I added a .vim subdirectory with files to detect .mu4 files (and git
commits, if you use git) and to set up a few niceties for editing. If
you run the "install-vim" Makefile target, it will tar up the .vim
directory and untar it in your home directory. Make sure it won't
overwrite anything before doing this; but truly the only possibility is
that it'll overwrite a .vim/ftdetect/git-commit.vim, but that's
unlikely.

I hope these make editing muFORTH files using Vim more pleasant.