Changed license (COPYRIGHT file) from MIT X to two-clause BSD
Why? Well, it was a toss-up in the first place. Being from the West
Coast (and in fact born in Berkeley!) I wanted to lean toward the BSD
license, but chose the MIT X license instead because it seemed more
explicit about what exactly was allowed.

Now, because it is clear about putting the copyright notice in the
documentation (in the case of a binary distribution) I'm going with BSD
instead. I think they are basically equivalent, anyway.