Tweaks to tty raw mode: now there are target and user submodes
I wanted to have a timeout on the serial port, so if there is nothing I
don't wait forever. But I don't want that for the human user,
interacting with a UI (such as the memory dumper/decompiler in
lib/see.mu4).

This change splits things up nicely, but still doesn't explain the
issues with the HC908 bootloader, which is why I'm undertaking this...