Initial support for Freescale MC908 bootloader
I had to make some small mods to the host serial support, and to the
memory dump code as well. Both are now temporarily hardcoded to work
with the 908. The serial port, when put into target-raw mode, sets up
two stop bits, and the memory dump only shows ASCII and hex bytes - in
part because it is re-reading the memory over and over to show each
line. I worked briefly on some ways to cache memory, and may add them
back, but it's not a priority. This chip is only a stepping-stone to the
S08 series anyway.

I had some fun with the code that reads memory. The 908 bootloader
presents the host with four commands for reading and writing memory:
Read, Write, Iread, and Iwrite. Read and Write both take a two-byte
address, and move a byte of data. Iread and Iwrite (the I is for
indexed) do postincrement reads and writes, starting from the last
transaction +1. If you think about this, it makes it hard to read an
even number of bytes, starting at an arbitrary address.

So I wrote two versions of simple caching - one caches two bytes; the
other, one - and then rewrote them, and intertwined their code so that
the differences and similarities would emerge. It took a long time, but
it was quite illuminating.

I haven't got to writing memory yet...