Snapshot: first successful BDM operation!
The code is still ugly, and it's still hard to do anything, and the 908
bootstrap still has occasional weird buggy behaviour - mostly not
setting the PC (on the stack) properly.

But since I was able to get something to work, I thought I'd snapshot
this as a moment in history.

The command I used was:

  ./muforth -d Bdm -d ExtClock -f target/HC08/loader.mu4

You then have to type:

  __outside
  chat

and then, to run the code:

  run

This stops to allow you to reset (power cycle) the S08 chip _by hand_.
After doing that, type

  go

to continue. This should return C8 in reg A, which is the contents of
the BDCSCR (Background debug controller status and control register!),
showing that it is in Active BDM, and using the 4M bus clock.