Got the 908's SCI working - but no luck connecting it to BDM
I tried a simple setup that essentially echoes characters between the
two interfaces, but the serial port is too slow. The BDM interface
(there should be a good term for this) times out after 512 (target) BDC
cycles, and gives up, aborting any command in progress. While I was able
to read the status "locally" and get valid results, when I try to do it
over SCI I get FF.

So I need a slightly more complex approach that can buffer a few
characters in each direction, so it can run the BDM commands at full
local speed. That shouldn't be too hard, but it's not as dead simple as
what I've got.