Added an elegant way to temporarily rebind global variables
Inspired by Klaus Schleisiek's elegant "uses and abuses of the return
stack" article, from FORML 1984, I've (mis)used my existing catch-throw
mechanism to add a similar feature to muFORTH.

I tried to hunt down any uses of the form

  <global> @ push   <do something>  pop <global> !

and change them to

  <global> preserve   <do something>

The new "preserve" has the nice feature that the global's value is reset
to the saved value either on a normal return from the word, or as a
result of an exception being thrown. This wasn't the case for many of
the ad hoc usages I found.