Improved stack behaviour during exceptions
It used to be that the stack got reset, both in mu_catch - which would
reset SP to what is what when the trapframe was constructed - and in
error (if the exception propagated that far), which would call sp!.

Now, neither of these things happens. This makes interactive use of the
system much nicer. Small typing or spelling mistakes don't blow away
carefully crafted computations on the stack. It's easy to retry what
failed with the same parameters. Stuff like that.

I also improved the stack over- and under-flow handling. Underflow
resets the stack (which is empty anyway), but overflow does not. The
overflow "error" is really a warning. The stack has 4096 slots, but has
32 cell over- and under-flow buffers at each end, so you'll get a
warning after pushing 4032 values onto the stack, but it's really not
going to overflow until 32 more values get pushed.

I added the word "drops" - which takes a count on the stack - to drop
multiple items. It can be a bit funny to use if the stack is full
though. Just typing

  200 drops

won't work. Pushing 200 will cause the error abort, and drops won't get
executed - but the 200 will be on the stack. So you really have to do

  200<Return>

and then

  drops<Return>

to make it happen, but this will work.

Using "sp!" interactively is an easy way to empty the entire stack.

I'm hoping these changes will be a major improvement in usability.