Errors no longer try to reset to "canonical" state
Like emptying the stack, this behaviour is rather annoying. While it
makes sense to want to return to some kind of "reasonable" state after
an error, jumping out of the meta-compiler, switching from hex back to
decimal and from compile to interpret state, and clearing the stack is
not only overkill but irritating. I don't want the system to jump
entirely out of what I was doing just because I made a small mistake.

This commit removes "canonical" and the call to it in "warm". It also
reuses code for patching of the compiler's "consumer" routine and
prompt.