Starting to deprecate "push" and "pop" in favor of >r and r>
As muFORTH moves steadily backwards towards fig-FORTH, I realise that
some of the decisions by "others" that I accepted as Obvious Goods, are
not necessarily such. (Clearing the data stack on error is one such
decision, which I have since revoked.) Using "push" as the name of the
word that pops a value from the data stack and pushes it on the return
stack, and "pop" as its inverse was an "innovation" that Chuck Moore
made years ago, perhaps in cmForth. It has stuck and is also a feature
of colorForth.

I thought it was a good idea, for rather a long time. Recently it has
seemed an odd choice, esp considering that in both cases you're pushing
one stack and popping the other. So why is one called "push" and the
other "pop"?

The old names - from the deep history of Forth - are ">r" ("to r") and
"r>" ("r from"). These graphically depict the movement of data from one
stack to the other. I'm starting to like them again.

Therefore, I made some changes in muFORTH. It's still possible to use
push and pop, and in fact tons of my code still does - I haven't changed
any of it. What I *have* done is to prepare the way for the Official
Deprecation, if not Outright Removal, of push and pop.

By hacking on the Makefile and the sed script that automagically
generates names from C functions, I was able to define >r and r> as the
true primitives. push and pop are synonyms, but you only get them if you
define CHUMOO in the Makefile. Currently this is done by default, but
may not be in the future.