Two tiny Makefile improvements - .asm and ${VERSOBJS}
Added .asm to SUFFIXES, so it's easy to get .asm versions of the .c
files.

Also, version.h used to depend on ${VERSOBJS:.o=.c} - the .c file names
for the list of files in VERSOBJS. This is silly. It works fine to
depend on the .o - because it depends on the .c file! So it will be
newer than version.h any time its corresponding .c file changes.