Successfully flashed a 908 part!
This checkin snapshots the code used to successfully write flash, plus
the initial breaking out of new simple chat code - to replace the slow
and awful MON08 interface - for the 908.

My efforts to relay between SCI and BDM were unsuccessful - I
essentially ran out of RAM. So I decided to go for the "let's figure out
how to flash the 908" route instead. It's easy! It worked the first
time.

My new chat code is pretty awesome too.

Next is to break apart (modularise) the boot code.