Modularised chat code; got basic Flash erase working
It's now much easier to load versions of the "chat" code to target
different implementations on the target. The initial one to target is of
course the 908's built-in MON08 - but it is awful and terrible, so we
compile a really simple & nice chat protocol engine - called simply
"chat", and use the MON08 interface to load it. (The load file
target/HC08/load-mon08.mu4 will do the trick.) Say "chat" and then "run"
and new code should be running.

Jumper the serial Rx and Tx over the SCI port, and say hello to chat!

You can now quit muforth and reload using target/HC08/load-908flash.mu4.
This will load only support for the "chat" protocol, and also adds
support for the ROM-based flash programming and erasing routines - which
work great!

Now I just have to put my "chat" program into flash, make sure I deal
with the watchdog, and when all seems perfect, write the reset vector.

One trouble with the 908 parts is that once you touch the reset vector
it's impossible to bring up the monitor again. You need a weird voltage
(>9v) and a weird frequency (9.302M or something). Bogus. So I'm going
to play around for as long as possible without touching the reset
vector.

This verison of chat also runs using the on-chip crystal oscillator. I
was using an external osc before. So now it's self-contained.