Renamed a few things, mostly; got rid of MON08 references
I was calling the built-in ROM monitor (on the 908 parts) "MON08", but I
think that's a P&E trademark. So now I just say "ROM monitor" or "908
ROM monitor".

I renamed the initial two load files to better indicate their purpose.
They are now called

  load-908boot0.mu4   and
  load-908boot1.mu4

I'm probably not going to try burning a Reset vector... the consequences
of getting it wrong are too high! And once the ("byteloader" chat) is
running in Flash, the chip rarely needs resetting. So it shouldn't be
too bad.

Onward to BDM!