"byteloader" protocol no longer uselessly echoes
While investigating slash debugging why the serial port was running 20x
slower than it should (I was seeing 500chars/s instead of 11k or so) I
decided to change my chat protocol so that it no longer sends back bytes
unless they mean something - like a byte read from memory, or the value
of the H or X register.

Having done this, I can now _write_ to memory at full wire speed - the
host plows ahead and sends WriteByte commands, and doesn't wait for echo
- and these get happily buffered and sent.

But reading memory is still really slow. The host sends the ReadByte
command, and waits. And eventually gets a response... but it takes a
"while". Certainly longer than the two hundred microsecs or so it _should_
take to send and receive a char at 115200 baud!

I have a theory. And it brings up the whole question of "UART FIFOs
considered harmful" as well. My theory is that my character-in-waiting
goes into the FIFO, and waits. Eventually the kernel (this is a Linux
box, as it happens - running 2.6.29) decides to flush the FIFO out and
sends a magic command to the hardware. The trouble is, I need to be able
to do this too. I tried the _termios_ commands tcflush() and tcsetattr
(with an "action" param of TCSADRAIN) but to no avail. I'm not sure I
have any options here.

It's not the end of the world - writing is fast, so writing buffers to
RAM to flash the chip will be fast. And interaction - like dumping
memory - "feels" fast enough. It's just annoying.

One other thing: this checkin makes possible a further checkin that will
make the chat protocol much more robust - by flushing the tty input
before each command, to make sure that things don't get out of sync. One
nice feature of the protocol - why I call it the byteloader protocol -
is that the target is only ever in one of two states: waiting for a
command byte, or waiting for _one_ data byte. It is never in the middle
of sending a long sequence of bytes, or waiting to read such a sequence.

This means that by flushing the host input queue, and then sending _one_
Sync command (which is no longer echoed) the target is guaranteed,
afterwards, to be in a known state: waiting for a command byte.

This is the code that I want to add next.