Periodically resynchronise the chat protocol
Whenever the host either sets the target HX (via SetHX ;-) or reads the
SP (via GetSP) it resynchronises, by flushing unread input and sending a
Sync character.

This is guaranteed to re-align the protocol. By doing it periodically
and automatically like this things _can_ get out of whack, but don't
stay that way, and important things - like writing code to a RAM buffer
to stage it for writing to Flash - always call SetHX _first_, so they
should never go wrong.

I "inserted" some noise into the stream and wasn't able to confuse .regs
or du!