Small improvements to 908 chat code; ready for Reset vector!
I wanted, after all, to try my luck burning a version of the byteloader
with a programmed reset vector. It's _much_ more convenient.

The downside is that it's possible (easy?) to mess things up, and when
that happens - it didn't initially, but I wanted to make a second small
change, and _that_ one caused my chip to get screwed up - it's hard to
recover. Once the reset vector is programmed you need 9.1 volts and a
9.8304M oscillator! (Who makes these things up?) I could get a nice high
voltage off my max232, but I couldn't figure out a way to get 9.8304M
out of any of my oscillators. I've got an 18.432M xtal; turns out they
are in a 15:8 ratio! Now go build me a circuit!

So, I tagged that chip and left it for dead. The second one is running
fine.

BTW: the little detail I wanted to add - that cost me that other chip -
was to no longer _assume_ I could just put something into the SCI data
xmit register! As I was debugging the BDM code - using SWIs!! - the byte
that the SWI handler wants to write was clobbering the BDM data I should
have been seeing. It was on the stack, but it wasn't getting sent. I got
suspicious, took out the SWIs, and it worked. After thinking about it, I
figured out that the SWI was overwriting the valid data before it was
able to leave the chip.

That's now fixed.