BDM code that actually works!
I finally came up with a reasonably elegant way to "script" the BDM
interaction from the host, while keeping the tight timing constraints.

A bit like the chatty byteloader, there is a small number of commands:
to escape back to the byteloader; to set how many bytes to write to
write to BDM, followed by the data; and to set how many bytes are
expected back from the BDM. It's quite nice, and all the BDM code takes
just under 256 bytes!