.regs now prints the "application" SP, rather than the SWI frame
I got confused debugging the BDM code, which uses a small buffer and
some counters on the stack. When I printed the registers it told me SP
was pointing somewhere odd - until I realised I was looking at the SWI
frame!

Now you don't see that, but if you subtract 6 from the SP shown and dump
memory, you'll see the SWI frame. So it's there if you really want it.
;-)