Working BDM implementation
The BDM "plumbing" - implemented on the 908 - is "scripted" by code
running on the host PC.

One issue with this: in my original code the Expect0 operation - which
posts the bytes from SCI to BDM and expects no BDM response - returned
nothing to the host, so it blithely sent _more_ bytes down the serial
line... but because my simple implementation on the 908 polls the SCI
interface, while it's away doing BDM things characters come in and are
lost.

So, I had to make a small hack: Expect0 expects nothing from BDM, but
the 908-side BDM code returns a dummy byte to the host, and the host
waits for this byte - so it doesn't continue until it knows that the BDM
transaction was completed.

With this change in place, and the hooks into "du", I was able to dump
the S08 memory and look inside my S08QG8 for the first time! It was
awesome, but it was too late to drink a Scotch. Maybe I'll celebrate
tonight instead...