Actually solved the filename/lineno on error issue
That first try was, well ... wrong. It wasn't low-level enough. Now any
time you throw() the right thing happens.

Another, collateral, change is that I only call longjmp with an integer
value. Before I was _casting_ a (char *) to an int and passing that;
this will break nicely on 64-bit machines... But no longer! Now I
capture info relevant at the moment of the throw - the filename and
message string - and then simply throw a non-zero int (-1, of course -
this is Forth!).

So, errors are much friendlier now, and the compiler is happier.