Made SciTx a subroutine in 908chat
I wanted it to be callable from other code.

However, this makes the chat loader 3 bytes longer, and now it buts up
against the control-block for the 908 flash routines... any more changes
to this code have to be done extra-carefully.

To make sure that no one gets bit by this, I put essentially an assert
that the first chunk of code hasn't crept into the flash control block.
That it turn meant breaking out the flash defs as a separate file...