Added a simple way to delete words from dictionary
We simply write a blank into the first character of the name. find is
rename to (find) and modified to return both the name and code fields. A
new word, find (like the old one) is added that calls (find) and throws
away the name part.

words is modified as well to skip the listing of deleted words.