Added Sync command to BDM; prefixed chat and BDM plumbing commands
I had been using "Sync" to refer to sending a no-op byte in the chat or
BDM protocol. I now call that "Idle". Sync refers to doing a BDM sync
pulse to measure the target's BDC clock.

In the process of doing this, I decided it made sense to put short
prefixes (c. and b.) on the low-level commands that do protocol things;
this way I don't accidentally do a chat command while the BDM protocol
is running.

I thought of combining the two... and may still. But I like that the
chat loader ("byteloader") is trivially small and simple. And stable.
BDM could still change, so tying them together seems like maybe a
mistake.