Getting read to compile chat byteloader for S08
This ended up involving a bunch of little changes that seem to have to
have nothing to do with 908 vs S08, though.

The obvious one that _does_ is that the chat protocol now includes an
"Id" command, which returns the low byte of the S08 id register (SDIDL)
on an S08, and FF on a 908.

This pushed the chat command loop into the flash control block, so I had
to move stuff around there, and decided to only allow flashing of 16
bytes at a time (instead of 32) to make a bit more room for the
byteloader code. Esp when initially bootstrapping this is important,
since the ROM monitor will reset the S (stack pointer) to 100, and then
do an SWI, so it's sitting at FA while we're doing our work. The code,
prior to changing from 32 byte flash buffer to 16, was sprawling up to
F2 or so, and I was getting nervous.

The device files were updated to use @ram @flash #ram and #flash for the
origin and length of ram and flash regions. Load files had to be changed
as well.

I think that's about it!