Baud rates for tty's can be set as low as 4800 now
In the process of recovering one of the 908's that I had "bricked" by
messing up the vectors and clobbering the clock trim byte (which lives
in the same flash page as the vectors), I managed to build a circuit
that gets 908s into the ROM monitor - except that since I lacked a
9.8304M crystal (who has even _heard_ of such a thing?) but had 4.9152M
ones (exactly half!) I thought "Hmm - maybe it'll work at 4800 instead
of 9600"!

And indeed it did. I was able to recover one of the two chips I had
screwed up. I think I fried the other one - the 9.1v pin was at more
like 9.4v (powered by our friend the max232). So I tapped off a 6.2v
zener in series with an LED. Now it's at 7.8v or so, which is better.
(It needs to be 2.5v higher than Vdd, which is 5v.)