BDM "chat" protocol is now self-contained
It is compiled at the beginning of flash, with a forward jump to the
"chat" loop. This way it's easy to find.

Also, the protocol now includes commands to set 4M or 8M bit clocks, to
drive the BKGD pin high, low, or tristate it, and to request a sync
pulse from the target.

Getting close to putting some code onto an S08 part!