S08 chat almost working!
I was able to have a conversation of sorts over SCI with the S08QG8!

It mostly worked - the first time I loaded the code! But I'm getting
weird spurious D8, E8, and F8 responses - esp for HX and for the
contents of flash. I'm not sure if I haven't tuned the oscillator close
enough to the baud rate, or if something more sinister is going on. It's
weird how well it works - dumping of registers works, dumping of memory
mostly does, except when the values become D8 E8 and F8 - but I can't
write memory. I tried tuning the trim a little bit but writing to it
didn't work.

Still, it's quite an accomplishment to have got to this point! Woohoo!