Got SCI-based chat working on S08
It was important that I successfully (ie, by not trying to do a bclr on
SOPT1) turn off the watchdog! Initially it seemed to work, but I would
sometimes get funny bytes back. It's amazing it worked at all - the
watchdog timer was firing and resetting the part over and over!

With that fixed it works fine. Now to try writing the S08 flash!