Version two of BDM protocol
As an experiment, I re-imagined the BDM-over-SCI protocol to be more
like the chat "byteloader" protocol. Now the BDM "host" is never
expecting  more than one byte over SCI, nor does it ever send more than
one byte over SCI. This yields the nice property that it's easy to
resynchronise the protocol.

I'm going to "align" the chat and bdm-over-sci protocols so the by
sending one Idle command and then an "exit back to chat" command (which
will be a no-op if we're already in the chat protocol) it will be easy
to get back and forth between the two modes (BDM and chat). Currently
this is a bit messy.

This conversion - to a similar "byte at a time" mode - is a start towards
that convergence.