Cleanup and minor refactoring of BDM SCI code
Code sharing between 908 and S08 versions is more explicit, and simpler.
Also I removed the dependency on chat's SciRx, so the load files
changed, and the way we compile macros had to change (so it gets the
assembler's versions of begin/if/while etc, rather than the Forth
compiler's).

I'm about to try burning the S08 BDM code into flash, but need to fix up
my flash code to work with chat - right now it's hardwired to burn the
flash via BDM only. So that's next...