BDM and chat both set up target for programming Flash
At reset the chat code trims the oscillator, turns off the watchdog, and
sets FCDIV (the flash clock divisor) to match its bus clock.

Now the BDM host code does the same thing - but for a different clock
rate. The chat code runs at 9.216M, in order to generate 115200 bps for
SCI. The BDM target - at least the QG chips - come out of reset and into
Active BDM mode with an untrimmed clock, and a bus clock of ~4M.

So, when we connect via BDM we use the factory trim to trim the clock to
4M and set FCDIV accordingly. We also turn off the watchdog and set the
initial SP to point to the end of RAM. Setting SP is important because
we use a buffer on the stack when programming the Flash.

I still need to make a generic layer so that the Flash code can do its
magic either via BDM or chat - still the only way to write to Flash is
via BDM. That's next.