Added a simple interface layer for accessing target regs and memory
Now I can flash the S08 via BDM or chat!

Incidentally, by moving the abstraction of indexed memory access to a
higher level - so that it is shared by chat, chat-monitor, and BDM - I
was able to simplify the code in several places. The host code that
talks to the 908 monitor is _much_ simpler, and seems to work better
than my fancy caching code. This sharing also means that the host code
only worries about setting its idea of HX, and whether it matches the
target's idea of it, in one place.

Because BDM allows access to the registers directly, whereas both chat
command sets read and write the register values that are sitting in an
SWI frame, I had to somehow integrate these different ideas of host-side
and target-side registers. The abstraction I ended up with is simple and
seems to work well. Basically, the keeps the registers in a buffer that
exactly mirrors an SWI frame. For the chat protocols, the register
values are copied back and forth simply by copying 6 bytes in a row. For
BDM the register values are extracted from the frame and put into the
target's registers, or read from the registers and put into the frame,
by using BDM commands. In every case, we have to be careful to read the
registers when we first connect to the target, and every time it
finishes executing a command; and to write them before launching it on
its way.

This version is not seamless and needs a bit of cleanup, but I want to
snapshot it anyway.