Changed `latest' to `last' and moved code out of dict.c
latest was used to capture the lfa of the latest word defined, so that
if `does>' was later executed, it could find the word and fix up its
parameter field with the new does IP.

This was silly and complicated, and done in C. I was also capturing this
info for -every- word defined; it only -needs- to be captured for words
that can later have their does pointer rewritten by `does>'; these words
are created by `constant', so I moved the code there (into startup.mu4).

Also, since I'm capturing this info specifically to rewrite the does
pointer, instead of storing the lfa and then having to calculate a pfa
from that, I capture pfa directly.

I've also changed the name of the variable - from `latest' to `last' -
to reflect previous usage, both in dforth, and in zenFORTH and
polyForth.