WIP - initial attempt at chaining vocabs
It works, and I finally understand why fig-forth had that funny-looking
`A081' hex literal - it's a fake name field, so that searches will chain
past it and never match (because it's a name consisting of a single
space character - and it's rather hard to create a token like that when
`token' drops leading and trailing whitespace!), but will allow
"dynamic" chaining.

A couple of terms. "Static" chaining is where one vocab links to a
specific word in another vocab, and this linking never changes. It's
useful for only using a subset of another vocab. "Dynamic" chaining is
where you link not to a -word- but to the vocab -itself-. In this case,
the vocab link pointer needs to "look like" a real header entry in the
vocab - hence the "fake name" field. Dynamic chaining is useful when you
always want to search the entirety of the linked-to vocab, even if it
grows after you've established the link to it.

This commit needs some work. It has endianness dependencies, and the C
code in dict.c doesn't properly represent the .forth. and .compiler.
vocabs with fake name fields.

But it proved the point, and it was an easy experiment.