Restructured dictionary names
Now it's possible to go from a name to a code field address - and vice
versa!

Names now precede link fields in the dictionary. The form of vocab heads
changed (for the better, mostly). Because of the way than names are
stored it's possible to go "backwards" from a code field address to
name. Names are stored with prefix padding, and suffixed by a length
byte, and names directly -precede- the link field.

This all adds up to wondrousness.